# LSM: Coalesce tables in memory