@sentientwaffle 2024-12-20
This fixes https://github.com/tigerbeetle/tigerbeetle/issues/2469.