@matklad 2024-12-20
This looks awesome! Everything looks right after the cursory look, but I'll do a careful review tomorrow! If we are releasing on Monday, I'd love to get this in, as it can save a lot of trouble for some users (but, of course, the users probably should arrange for sending us bigger batches)!