@sentientwaffle 2024-12-22
Ah, the value-count-max vs batch-limit is a little trickier than I first thought. :thinking: